test(workflow): the trigger fixture outlives the suite it was written in
Moving the dispatch benchmark behind `integration && bench` took two
declarations with it that a suite on the plain `integration` tag had started
using while this branch was open: leadCreatedEventType and
seedAllStarterAutomations. The replay suite then failed to COMPILE in the lane
that runs on every push, which is worse than a failing test — the shard reports
a build error and no test in the package runs at all.

They live in their own fixture now, on the tag both builds see. The
bench-only helpers stay where they are: an envelope builder and a lead seeder
that only a measurement needs.

The general shape, written into the fixture: a helper shared across a build-tag
boundary belongs on the weaker tag. Left in the benchmark it is invisible to
the build that matters, and nothing about the benchmark says a merge-gate suite
depends on it.
